Gradebook App
With the syllabus that was used for this class, I created a program that could calculate the grade you need on the final exam in order to get a grade that you would prefer. The output is the grade required on the final exam to receive the grade you are hopeful to get with the weights for that specific class.